@@ -11,40 +11,60 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.0.3.RELEASE</version>
-        <!--        <version>2.4.3.RELEASE</version>-->
+        <version>3.0.2</version>
+        <relativePath/> <!-- lookup parent from repository -->
     </parent>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
-        <spring.boot.version>2.0.3.RELEASE</spring.boot.version>
-        <!--<sprint.version>5.0.7.RELEASE</sprint.version>-->
-        <spring.version>5.3.22</spring.version>
+        <java.version>17</java.version>
 
         <image.repo.prefix>swr.cn-north-4.myhuaweicloud.com/synctech</image.repo.prefix>
+        <image.basic>openjdk:17-jdk-buster</image.basic>
         <image.repo.auth.username>cn-north-4@A0AAZNO5HZN1RZ5TWT9W</image.repo.auth.username>
-        <image.repo.auth.password>15b9549fe73f98a846d5453fd8df0ea0b45c3cc3b9d4ffe31aaa9ad5b90877b8</image.repo.auth.password>
+        <image.repo.auth.password>
+            15b9549fe73f98a846d5453fd8df0ea0b45c3cc3b9d4ffe31aaa9ad5b90877b8
+        </image.repo.auth.password>
+        <maven.build.timestamp.format>yyyyMMdd.HHmm</maven.build.timestamp.format>
     </properties>
 
     <modules>
-        <module>component</module>
+        <module>redis</module>
+        <!-- <module>minio</module> -->
         <module>common</module>
         <module>task</module>
         <module>ledger</module>
         <module>manager</module>
-        <module>administrator</module>
         <module>mining-pool</module>
-        <module>mockserver</module>
     </modules>
+
+    <repositories>
+        <repository>
+            <id>spring-milestone</id>
+            <name>Spring milestone</name>
+            <url>https://repo.spring.io/milestone</url>
+            <snapshots>
+                <enabled>false</enabled>
+            </snapshots>
+        </repository>
+    </repositories>
+    <pluginRepositories>
+        <pluginRepository>
+            <id>spring-milestone</id>
+            <name>Spring milestone</name>
+            <url>https://repo.spring.io/milestone</url>
+            <snapshots>
+                <enabled>false</enabled>
+            </snapshots>
+        </pluginRepository>
+    </pluginRepositories>
 
     <dependencies>
         <!-- test -->
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <!--<version>4.12</version>-->
             <version>4.13.2</version>
             <scope>test</scope>
         </dependency>
