@@ -1,7 +1,8 @@
 package com.synctech.statter.base.entity;
 
+import com.alibaba.fastjson2.annotation.JSONField;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -24,60 +25,82 @@ public class Miner implements Serializable {
     public static final long TWO_MINUTE_MILLIS = 2 * 60 * 1000;
     public static final long TEN_MINUTE_MILLIS = 10 * 60 * 1000;
 
-    @ApiModelProperty(name = "sn", value = "Mining machine SN code, it's a string with a length of 32 by default.")
+    @Schema(name = "sn", description = "Mining machine SN code, it's a string with a length of 32 by default.")
     @JsonProperty("sn")
     String sn;
+
+    int status;
+
+    /**
+     * The time of leave factory
+     */
+    @Schema(name = "leaveFactory")
+    @JsonProperty("leaveFactory")
+    @JSONField(format = "yyyy-MM-dd HH:mm:ss")
+    Timestamp leaveFactory;
 
     /**
      * @see Wallet#address
      */
-    @ApiModelProperty(name = "wa", value = "The wallet address bundled by the mining machine, defaults to the extension string of length " + WALLET_ADDRESS_LENGTH + ", starting with ST.")
+    @Schema(name = "wa", description = "The wallet address bundled by the mining machine, defaults to the extension string of length " + WALLET_ADDRESS_LENGTH + ", starting with ST.")
     @JsonProperty("wa")
     String walletAddress;
 
     /**
      * @see Promotion#address
      */
-    @ApiModelProperty(name = "pa", value = "The ore pond wallet address corresponding to the wallet bundled by the mining machine, defaults to the extended string of length " + WALLET_ADDRESS_LENGTH + ", and ST starts.")
+    @Schema(name = "pa", description = "The ore pond wallet address corresponding to the wallet bundled by the mining machine, defaults to the extended string of length " + WALLET_ADDRESS_LENGTH + ", and ST starts.")
     @JsonProperty("pa")
     String promotionAddress;
 
-    @ApiModelProperty(name = "bd", value = "The date of the miner bind to wallet")
+    @Schema(name = "bd", description = "The date of the miner bind to wallet")
     @JsonProperty("bd")
+    @JSONField(format = "yyyy-MM-dd HH:mm:ss")
     Timestamp bindDate;
 
-    @ApiModelProperty(name = "hp", value = "Whether the miner pledge has been completed")
+    @Schema(name = "hp", description = "Whether the miner pledge has been completed")
     @JsonProperty("hp")
     boolean hasPledged;
 
     /**
      * @see Process
      */
-    @ApiModelProperty(name = "ppi", value = "The pledge process ID is undergoing or ended.")
+    @Schema(name = "ppi", description = "The pledge process ID is undergoing or ended.")
     @JsonProperty("ppi")
     long pledgeProcessId;
 
-    @ApiModelProperty(name = "ht", value = "Whether the mine tax has been completed.")
+    @Schema(name = "ht", description = "Whether the mine tax has been completed.")
     @JsonProperty("ht")
     boolean hasTaxed;
 
     /**
      * @see Process
      */
-    @ApiModelProperty(name = "tpi", value = "The tax process ID is undergoing or ended.")
+    @Schema(name = "tpi", description = "The tax process ID is undergoing or ended.")
     @JsonProperty("tpi")
     long taxProcessId;
 
-    @ApiModelProperty(name = "mhh", value = "The maximum hash of this mining machine.")
+    @Schema(name = "mhh", description = "The maximum hash of this mining machine.")
     @JsonProperty("mhh")
     long maxHistoryHash;
 
-    @ApiModelProperty(name = "mi", value = "Id generated by the configuration parameter according to the physics equipment of the mining machine")
+    @Schema(name = "mi", description = "Id generated by the configuration parameter according to the physics equipment of the mining machine")
     @JsonProperty("mi")
     String machineId;
 
-    @ApiModelProperty(name = "cpuModuleName", value = "The model of the mining machine's cpu")
+    @Schema(name = "cpuModuleName", description = "The model of the mining machine's cpu")
     @JsonProperty("cpuModuleName")
     String cpuModuleName;
+
+    @Schema(name = "v", description = "")
+    @JsonProperty("v")
+    int v;
+    @Schema(name = "ver", description = "")
+    @JsonProperty("ver")
+    String ver;
+
+    @Schema(name = "dv", description = "")
+    @JsonProperty("dv")
+    String dv;
 
 }
