@@ -1,7 +1,7 @@
 package com.synctech.statter.base.entity;
 
 import com.alibaba.fastjson.JSONObject;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -22,39 +22,40 @@ public class Promotion implements Serializable {
 
     Timestamp updatedTime;
 
-    byte status;
+    int status;
+    int visible;
 
     /**
      * @see Wallet#address
      */
-    @ApiModelProperty(name = "address", value = "The income wallet address of the mining pool, defaults to the length " + WALLET_ADDRESS_LENGTH + ", the extensive string that has begun at the beginning of ST")
+    @Schema(name = "address", description = "The income wallet address of the mining pool, defaults to the length " + WALLET_ADDRESS_LENGTH + ", the extensive string that has begun at the beginning of ST")
     String address;
 
-    @ApiModelProperty(name = "minerCount", value = "Number of mining machines under the jurisdiction")
+    @Schema(name = "minerCount", description = "Number of mining machines under the jurisdiction")
     int minerCount;
 
-    @ApiModelProperty(value = "the statistics standard hash of the miners under the promotion")
+    @Schema(description = "the statistics standard hash of the miners under the promotion")
     long hash;
 
-    @ApiModelProperty(name = "code", value = "the unique code of the promotion")
+    @Schema(name = "code", description = "the unique code of the promotion")
     String code;
 
-    @ApiModelProperty(name = "alias", value = "alias name")
+    @Schema(name = "alias", description = "alias name")
     String alias;
 
-    @ApiModelProperty(name = "introduction", value = "The introduction about promotion")
+    @Schema(name = "introduction", description = "The introduction about promotion")
     String introduction;
 
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-    @ApiModelProperty(name = "manage secret key", value = "Used for asking or refreshing secret key")
+    @Schema(name = "manage secret key", description = "Used for asking or refreshing secret key")
     String managementKey;
-    @ApiModelProperty(name = "secret key", value = "Used for ask the api of the mining pool")
+    @Schema(name = "secret key", description = "Used for ask the api of the mining pool")
     String secretKey;
-    @ApiModelProperty(name = "secret key refresh time", value = "Secret key can only be refreshed once in 24 hours")
+    @Schema(name = "secret key refresh time", description = "Secret key can only be refreshed once in 24 hours")
     Timestamp secretKeyUptTime;
 
-//    @ApiModelProperty(name = "authority", value = "Extend column for expanding modules of the mining pool in future, it's useless now.")
+//    @Schema(name = "authority", description = "Extend column for expanding modules of the mining pool in future, it's useless now.")
 //    String authority;
 
     @Override
