@@ -17,7 +17,7 @@ public class Rule extends BaseEntity {
         MinerPledge((byte) 21),  // miner pledge // {"amount": "100.00","targetAddress":"ST60ff64f90e2400a86591f5db938ea4f3","pledgeMinDays": "180"}
         WalletPledge((byte) 31), // wallet tax/pledge  // {"amount": "10000.00","targetAddress":"ST60ff64f90e2400a86591f5db938ea4f3"}
 
-        WhiteList((byte) 51), // white list switch  // {"minerWhiteListSwitch": false}
+//        WhiteList((byte) 51), // white list switch  // {"minerWhiteListSwitch": false}
 
         MinerMachine((byte) 91), // mining machine(miner) // {"updateScriptUrl": "http://192.168.1.164/downloads/statter.update.v1.0.1.sh","version": "v1.0.1"}
         ;
