@@ -1,12 +1,13 @@
 package com.synctech.statter.base.entity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.sql.Timestamp;
 
 import static com.synctech.statter.constant.Constant.WALLET_ADDRESS_LENGTH;
 
@@ -15,30 +16,34 @@ import static com.synctech.statter.constant.Constant.WALLET_ADDRESS_LENGTH;
 @NoArgsConstructor
 public class Wallet implements Serializable {
 
-    @ApiModelProperty(name = "a", value = "Wallet Address, defaults to the extension string that has begun at the beginning of ST.")
+    @Schema(name = "a", description = "Wallet Address, defaults to the extension string that has begun at the beginning of ST.")
     @JsonProperty("a")
     String address;
 
     /**
      * @see Promotion#address
      */
-    @ApiModelProperty(name = "pa", value = "The wallet address belongs to the ore pond wallet address, the default is " + WALLET_ADDRESS_LENGTH + ", the laid -up string that has begun.")
+    @Schema(name = "pa", description = "The wallet address belongs to the ore pond wallet address, the default is " + WALLET_ADDRESS_LENGTH + ", the laid -up string that has begun.")
     @JsonProperty("pa")
     String promotionAddress;
 
-    @ApiModelProperty(name = "alias", value = "Alias name for this wallet account")
+    @Schema(name = "joinTime", description = "The time when the wallet joined the promotion.")
+    @JsonProperty("joinTime")
+    Timestamp joinTime;
+
+    @Schema(name = "alias", description = "Alias name for this wallet account")
     @JsonProperty("alias")
     String alias;
 
-    @ApiModelProperty(name = "hp", value = "Whether the wallet pledge has been completed")
+    @Schema(name = "hp", description = "Whether the wallet pledge has been completed")
     @JsonProperty("hp")
     boolean hasPledged;
 
-    @ApiModelProperty(name = "ppi", value = "The pledge process ID is undergoing or ending")
+    @Schema(name = "ppi", description = "The pledge process ID is undergoing or ending")
     @JsonProperty("ppi")
     long pledgeProcessId;
 
-    @ApiModelProperty(name = "minerCount", value = "Number of mining machines binding under this wallet address")
+    @Schema(name = "minerCount", description = "Number of mining machines binding under this wallet address")
     int minerCount;
 
 }
