@@ -1,5 +1,6 @@
 package com.synctech.statter.common.service.service;
 
+import cn.hutool.core.collection.CollectionUtil;
 import com.alibaba.fastjson.JSONObject;
 import com.synctech.statter.base.entity.Rule;
 import com.synctech.statter.base.entity.White;
@@ -7,23 +8,27 @@ import com.synctech.statter.base.mapper.WhiteMapper;
 import com.synctech.statter.constant.CacheKey;
 import com.synctech.statter.redis.jedis.JedisService;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
 
-import javax.annotation.Resource;
+import java.util.*;
+import java.util.stream.Collectors;
 
 @Slf4j
 @Service
 public class WhiteService {
 
-    @Resource
+    @Autowired
     JedisService jedisService;
 
-    @Resource
+    @Autowired
     WhiteMapper whiteMapper;
 
-    @Resource
-    RuleService ruleService;
+//    @Autowired
+//    RuleService ruleService;
 
     @Transactional
     public void add(String n, White.Type t) {
@@ -31,31 +36,38 @@ public class WhiteService {
         whiteMapper.add(w);
     }
 
-    public boolean isWhiteModelOpen() {
-        JSONObject j = null;
-        try {
-            j = ruleService.get(Rule.Type.WhiteList);
-            if (null == j) return false;
-        } catch (Exception e) {
-            log.warn("white list rule has not be setting");
-            return false;
-        }
-        return j.getBoolean("minerWhiteListSwitch");
-    }
+//    public Set<String> findAll(byte type) {
+//        List<White> list = whiteMapper.findByType(type);
+//        return CollectionUtil.isEmpty(list) ? new HashSet<>() : list.stream().map(w -> w.getN()).collect(Collectors.toSet());
+//    }
+
+//    public boolean isWhiteModelOpen() {
+//        JSONObject j = null;
+//        try {
+//            j = ruleService.get(Rule.Type.WhiteList);
+//            if (null == j) return false;
+//        } catch (Exception e) {
+//            log.warn("white list rule has not be setting");
+//            return false;
+//        }
+//        return j.getBoolean("minerWhiteListSwitch");
+//    }
 
     public boolean isWhiteMiner(String n) {
-        if (!this.isWhiteModelOpen()) return false;
-        return this.isWhiteMinerImpl(n);
+        return jedisService.hexists(CacheKey.CACHEKEY_ADMIN_WHITE_LIST_MINER, n);
+//        int exist = whiteMapper.exist(n, White.Type.Miner.getValue());
+//        return exist == 1;
     }
 
-    public boolean isWhiteMinerImpl(String n) {
-        Boolean exist = jedisService.hget(CacheKey.CACHEKEY_ADMIN_WHITE_LIST_MINER, n, Boolean.class);
-        if (null == exist) {
-            int we = whiteMapper.exist(n, White.Type.Miner.getValue());
-            exist = we > 0 ? true : false;
-            jedisService.hset(CacheKey.CACHEKEY_ADMIN_WHITE_LIST_MINER, n, exist);
-        }
-        return exist;
+    public void cacheWhiteMinerList() {
+        List<White> list = whiteMapper.findByType((byte) 1);
+        jedisService.del(CacheKey.CACHEKEY_ADMIN_WHITE_LIST_MINER);
+        if (CollectionUtils.isEmpty(list)) return;
+        Map<String, String> map = list.stream().collect(Collectors.toMap(White::getN, w -> "1"));
+        jedisService.hsetAll(CacheKey.CACHEKEY_ADMIN_WHITE_LIST_MINER, map);
+//        for (int i = 0; i < list.size(); i++) {
+//            jedisService.hset(CacheKey.CACHEKEY_ADMIN_WHITE_LIST_MINER, list.get(i).getN(), true);
+//        }
     }
 
 }
