@@ -1,6 +1,7 @@
 package com.synctech.statter.manager.api.controller.v1;
 
 import com.synctech.statter.base.entity.ApplyForPromotion;
+import com.synctech.statter.base.entity.Promotion;
 import com.synctech.statter.base.mapper.ApplyForPromotionMapper;
 import com.synctech.statter.common.service.service.PromotionService;
 import com.synctech.statter.common.service.service.WalletService;
@@ -8,33 +9,36 @@ import com.synctech.statter.common.service.vo.info.WalletVo;
 import com.synctech.statter.constant.HttpStatusExtend;
 import com.synctech.statter.constant.restful.AppBizException;
 import com.synctech.statter.constant.restful.DataResponse;
-import io.swagger.annotations.*;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.enums.ParameterIn;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import javax.annotation.Resource;
-
 @Slf4j
-@Api(value = "api about pool for dapp")
-@RequestMapping("v1/apply/prom")
+@Tag(name = "manager: apply promotion")
+@RequestMapping("statter/manager/api/v1/apply/prom")
 @RestController("applyPromotionController")
 public class ApplyPromotionController {
 
-    @Resource
+    @Autowired
     ApplyForPromotionMapper applyForPromotionMapper;
 
-    @Resource
+    @Autowired
     WalletService walletService;
 
-    @Resource
+    @Autowired
     PromotionService promotionService;
 
-    @ApiOperation(httpMethod = "GET", value = "get the apply for upgrading to a promotion")
-    @ApiResponses({@ApiResponse(code = 200, message = "OK", response = ApplyForPromotion.class)})
+    @Operation(method = "GET", description = "get the apply for upgrading to a promotion", parameters = @Parameter(name = "address", description = "wallet address", in = ParameterIn.PATH, schema = @Schema(implementation = String.class), required = true), responses = @ApiResponse(content = @Content(mediaType = "application/json", schema = @Schema(implementation = ApplyForPromotion.class))))
     @GetMapping("/{address}")
-    public String get(@ApiParam(name = "address", value = "wallet address", type = "String", required = true)
-                      @PathVariable("address") String address) {
+    public String get(@PathVariable("address") String address) {
         if (StringUtils.isBlank(address)) {
             return DataResponse.fail(HttpStatusExtend.ERROR_INVALID_REQUEST);
         }
@@ -45,12 +49,11 @@ public class ApplyPromotionController {
         return DataResponse.success(applyForPromotion);
     }
 
-    @ApiOperation(httpMethod = "POST", value = "commit the apply for upgrading to a promotion")
-    @ApiResponses({@ApiResponse(code = 200, message = "OK", response = String.class)})
+    @Operation(method = "POST", description = "commit the apply for upgrading to a promotion", requestBody = @io.swagger.v3.oas.annotations.parameters.RequestBody(description = "apply information for promotion", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ApplyForPromotion.class))))
     @PostMapping("")
-    public String apply(@ApiParam(name = "params", type = "ApplyForPromotion", required = true)
-                        @RequestBody() ApplyForPromotion body) {
-        if (null == body || StringUtils.isBlank(body.getAddress()) || StringUtils.isBlank(body.getAlias())) {
+    public String apply(@RequestBody() ApplyForPromotion body) {
+        if (null == body || StringUtils.isBlank(body.getAddress()) ||
+                StringUtils.isBlank(body.getAlias())) {
             return DataResponse.fail(HttpStatusExtend.ERROR_INVALID_REQUEST);
         }
         String address = body.getAddress();
@@ -58,16 +61,18 @@ public class ApplyPromotionController {
         ApplyForPromotion applyForPromotion = applyForPromotionMapper.find(address);
         if (null != applyForPromotion)
             return DataResponse.fail(new AppBizException(HttpStatusExtend.ERROR_WALLET_APPLY_FOR_PROMOTION_EXIST));
+        Promotion p = promotionService.findByAlias(body.getAlias());
+        if (null != p)
+            return DataResponse.fail(new AppBizException(HttpStatusExtend.ERROR_WALLET_ALIAS_FOR_PROMOTION_EXIST));
         applyForPromotionMapper.add(body);
         return DataResponse.success();
     }
 
-    @ApiOperation(httpMethod = "PUT", value = "edit the apply for upgrading to a promotion")
-    @ApiResponses({@ApiResponse(code = 200, message = "OK", response = String.class)})
+    @Operation(method = "PUT", description = "edit the apply for upgrading to a promotion", requestBody = @io.swagger.v3.oas.annotations.parameters.RequestBody(description = "apply information for promotion", content = @Content(mediaType = "application/json", schema = @Schema(implementation = ApplyForPromotion.class))))
     @PutMapping("")
-    public String edit(@ApiParam(name = "params", type = "ApplyForPromotion", required = true)
-                       @RequestBody() ApplyForPromotion body) {
-        if (null == body || StringUtils.isBlank(body.getAddress()) || StringUtils.isBlank(body.getAlias())) {
+    public String edit(@RequestBody() ApplyForPromotion body) {
+        if (null == body || StringUtils.isBlank(body.getAddress()) ||
+                StringUtils.isBlank(body.getAlias())) {
             return DataResponse.fail(HttpStatusExtend.ERROR_INVALID_REQUEST);
         }
         String address = body.getAddress();
@@ -77,8 +82,10 @@ public class ApplyPromotionController {
             return DataResponse.fail(new AppBizException(HttpStatusExtend.ERROR_WALLET_APPLY_FOR_PROMOTION_EXIST));
         else if (ApplyForPromotion.Status.PASS.compare(applyForPromotion.getStatus()))
             return DataResponse.fail(HttpStatusExtend.ERROR_INVALID_REQUEST);
-        if (StringUtils.isNotBlank(body.getAlias())) applyForPromotion.setAlias(body.getAlias());
-        if (StringUtils.isNotBlank(body.getIntroduction())) applyForPromotion.setIntroduction(body.getIntroduction());
+        if (StringUtils.isNotBlank(body.getAlias()))
+            applyForPromotion.setAlias(body.getAlias());
+        if (StringUtils.isNotBlank(body.getIntroduction()))
+            applyForPromotion.setIntroduction(body.getIntroduction());
         applyForPromotionMapper.updateInfo(body);
         return DataResponse.success();
     }
