@@ -4,28 +4,27 @@ import com.alibaba.fastjson.JSONObject;
 import com.synctech.statter.base.entity.Rule;
 import com.synctech.statter.common.service.service.RuleService;
 import com.synctech.statter.constant.restful.DataResponse;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.annotation.Resource;
-
 @Slf4j
-@Api(value = "api about biz rule")
-@RequestMapping("v1/rule")
+@Tag(name = "manager: rule")
+@RequestMapping("statter/manager/api/v1/rule")
 @RestController("openRuleController")
 public class RuleController {
 
-    @Resource
+    @Autowired
     RuleService ruleService;
 
-    @ApiOperation(httpMethod = "GET", value = "tax and pledge rules of miners ")
-    @ApiResponses({@ApiResponse(code = 200, message = "OK", response = String.class)})
+    @Operation(method = "GET", description = "tax and pledge rules of miners", responses = @ApiResponse(content = @Content(mediaType = "application/json", schema = @Schema(implementation = String.class))))
     @GetMapping("/miner/tap")
     public String minerTaxAndPledge() {
         JSONObject minerManageRule = new JSONObject();
@@ -36,8 +35,7 @@ public class RuleController {
         return DataResponse.success(minerManageRule);
     }
 
-    @ApiOperation(httpMethod = "GET", value = "tax or pledge rules of wallets")
-    @ApiResponses({@ApiResponse(code = 200, message = "OK", response = String.class)})
+    @Operation(method = "GET", description = "tax or pledge rules of wallets", responses = @ApiResponse(content = @Content(mediaType = "application/json", schema = @Schema(implementation = String.class))))
     @GetMapping("/wallet/p")
     public String walletPledge() {
         JSONObject minerManageRule = new JSONObject();
