@@ -1,6 +1,6 @@
 package com.synctech.statter.manager.api.controller.v1.vo;
 
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -16,22 +16,23 @@ import static com.synctech.statter.constant.Constant.WALLET_ADDRESS_LENGTH;
 @Accessors(chain = true)
 public class PromotionInfoVo implements Serializable {
 
-    @ApiModelProperty(name = "code", value = "the unique code of the promotion")
+    @Schema(name = "code", description = "the unique code of the promotion")
     String code;
 
-    @ApiModelProperty(name = "alias", value = "The alias name of the promotion")
+    @Schema(name = "alias", description = "The alias name of the promotion")
     String alias;
 
-    @ApiModelProperty(name = "address", value = "The income wallet address of the mining pool is default. The default is " + WALLET_ADDRESS_LENGTH + " and the extensive string that has begun.")
+    @Schema(name = "address", description = "The income wallet address of the mining pool is default. The default is "
+            + WALLET_ADDRESS_LENGTH + " and the extensive string that has begun.")
     String address;
 
-    @ApiModelProperty(name = "introduction", value = "The introduction about promotion")
+    @Schema(name = "introduction", description = "The introduction about promotion")
     String introduction;
 
-    @ApiModelProperty(name = "minerCount", value = "Number of mining machines under this mining pool.")
+    @Schema(name = "minerCount", description = "Number of mining machines under this mining pool.")
     int minerCount;
 
-    @ApiModelProperty(name = "hash", value = "the real-time hash under this mining pool")
+    @Schema(name = "hash", description = "the real-time hash under this mining pool")
     long hash;
 
 }
