@@ -10,39 +10,43 @@ import com.synctech.statter.constant.restful.DataResponse;
 import com.synctech.statter.mining.pool.api.controller.v1.CommonController;
 import com.synctech.statter.mining.pool.api.controller.v1.admin.vo.RefreshSecretKey;
 import com.synctech.statter.redis.jedis.JedisService;
-import io.swagger.annotations.*;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import javax.annotation.Resource;
-
 @Slf4j
-@Api(value = "api of manage secret key")
-@RequestMapping("v1/sk")
+@Tag(name = "mining pool: secret key")
+@RequestMapping("statter/mining/pool/api/v1/sk")
 @RestController()
 public class SecretKeyController extends CommonController {
 
-    @Resource
+    @Autowired
     JedisService jedisService;
 
-    @Resource
+    @Autowired
     PromotionService promotionService;
 
-    @ApiOperation(httpMethod = "POST", value = "Refresh secret key by the management key.")
-    @ApiResponses({@ApiResponse(code = 200, message = "OK", response = String.class)})
+    @Operation(method = "POST", description = "Refresh secret key by the management key.", requestBody = @io.swagger.v3.oas.annotations.parameters.RequestBody(content = @Content(mediaType = "application/json", schema = @Schema(implementation = RefreshSecretKey.Req.class))), responses = @ApiResponse(content = @Content(mediaType = "application/json", schema = @Schema(implementation = String.class))))
     @PostMapping("refresh")
-    public String refresh(@ApiParam(type = "json", required = true) @RequestBody RefreshSecretKey.Req req) {
+    public String refresh(@RequestBody RefreshSecretKey.Req req) {
         if (null == req || StringUtils.isBlank(req.getA()) || StringUtils.isBlank(req.getMk()))
             throw new AppBizException(HttpStatusExtend.ERROR_INVALID_REQUEST);
         Promotion p = jedisService.hget(CacheKey.CACHEKEY_INFO_PROMOTION_BY_ADDRESS, req.getA(), Promotion.class);
-        if (null == p) throw new AppBizException(HttpStatusExtend.ERROR_INVALID_REQUEST);
+        if (null == p)
+            throw new AppBizException(HttpStatusExtend.ERROR_INVALID_REQUEST);
         if (!StringUtils.equals(p.getManagementKey(), req.getMk())) {
             throw new AppBizException(HttpStatusExtend.ERROR_INVALID_REQUEST);
-        } else if (StringUtils.isNotBlank(p.getSecretKey()) && (System.currentTimeMillis() - p.getSecretKeyUptTime().getTime() < 86400000)) {
+        } else if (StringUtils.isNotBlank(p.getSecretKey())
+                && (System.currentTimeMillis() - p.getSecretKeyUptTime().getTime() < 86400000)) {
             return DataResponse.success(p.getSecretKey());
         }
         promotionService.refreshSecretKey(req.getA());
