@@ -1,6 +1,6 @@
 package com.synctech.statter.mining.pool.api.controller.v1.promotion.vo;
 
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
 import java.io.Serializable;
@@ -10,13 +10,13 @@ import static com.synctech.statter.constant.Constant.WALLET_ADDRESS_LENGTH;
 @Data
 public class PromotionSimpleVo implements Serializable {
 
-    @ApiModelProperty(name = "code", value = "the unique code of the promotion")
+    @Schema(name = "code", description = "the unique code of the promotion")
     String code;
 
-    @ApiModelProperty(name = "alias", value = "The alias name of the promotion")
+    @Schema(name = "alias", description = "The alias name of the promotion")
     String alias;
 
-    @ApiModelProperty(name = "introduction", value = "The introduction about promotion")
+    @Schema(name = "introduction", description = "The introduction about promotion")
     String introduction;
 
 }
