@@ -1,7 +1,7 @@
 package com.synctech.statter.mining.pool.api.controller.v1.promotion.vo;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
 import java.io.Serializable;
@@ -11,23 +11,24 @@ import static com.synctech.statter.constant.Constant.WALLET_ADDRESS_LENGTH;
 @Data
 public class WalletSimpleVo implements Serializable {
 
-    @ApiModelProperty(name = "a", value = "Wallet Address, defaults to the extension string that has begun at the beginning of ST.")
+    @Schema(name = "a", description = "Wallet Address, defaults to the extension string that has begun at the beginning of ST.")
     @JsonProperty("a")
     String address;
 
-    @ApiModelProperty(name = "pa", value = "The wallet address belongs to the ore pond wallet address, the default is " + WALLET_ADDRESS_LENGTH + ", the laid -up string that has begun.")
+    @Schema(name = "pa", description = "The wallet address belongs to the ore pond wallet address, the default is "
+            + WALLET_ADDRESS_LENGTH + ", the laid -up string that has begun.")
     @JsonProperty("pa")
     String promotionAddress;
 
-    @ApiModelProperty(name = "alias", value = "Alias name for this wallet account")
+    @Schema(name = "alias", description = "Alias name for this wallet account")
     @JsonProperty("alias")
     String alias;
 
-    @ApiModelProperty(name = "hp", value = "Whether the wallet pledge has been completed")
+    @Schema(name = "hp", description = "Whether the wallet pledge has been completed")
     @JsonProperty("hp")
     boolean hasPledged;
 
-    @ApiModelProperty(name = "minerCount", value = "Number of mining machines binding under this wallet address")
+    @Schema(name = "minerCount", description = "Number of mining machines binding under this wallet address")
     int minerCount;
 
 }
