@@ -1,7 +1,7 @@
 package com.synctech.statter.mining.pool.api.controller.v1.wallet.vo;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.annotations.ApiModelProperty;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
@@ -13,11 +13,12 @@ public abstract class JoinPromotion {
     @Accessors(chain = true)
     public static class Req {
 
-        @ApiModelProperty(name = "a", value = "Wallet Address, defaults to the extension string that has begun at the beginning of ST.")
+        @Schema(name = "a", description = "Wallet Address, defaults to the extension string that has begun at the beginning of ST.")
         @JsonProperty("a")
         String address;
 
-        @ApiModelProperty(name = "pa", value = "The wallet address belongs to the ore pond wallet address, the default is " + WALLET_ADDRESS_LENGTH + ", the laid -up string that has begun.")
+        @Schema(name = "pa", description = "The wallet address belongs to the ore pond wallet address, the default is "
+                + WALLET_ADDRESS_LENGTH + ", the laid -up string that has begun.")
         @JsonProperty("pa")
         String promotionAddress;
 
@@ -28,6 +29,5 @@ public abstract class JoinPromotion {
     public static class Resp {
 
     }
-
 
 }
