@@ -6,7 +6,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import javax.annotation.Resource;
+import org.springframework.beans.factory.annotation.Autowired;
 
 @Component
 @Slf4j
@@ -15,7 +15,7 @@ public class WalletJobs {
     @Autowired
     JedisService jedisService;
 
-    @Resource
+    @Autowired
     MinerMapper minerMapper;
 
 }
